refactor form into separate component
